refactor: endpoint 재구성
- 데이터 조작 관련 api 들은 admin 라우터로 분리
- 계층별 조회 api endpoint 수정(levels -> hierarchy)
